+++ aaran/Transaction/Models/AccountBook.php
@@ -38,7 +38,17 @@ class AccountBook extends Model
     {
         return $this->belongsTo(Common::class);
 
+    }
 
+    public function transaction(): BelongsTo
+    {
+        return $this->belongsTo(Transaction::class,'account_book_id','id');
+
+    }
+
+    public function accountBook(): BelongsTo
+    {
+        return $this->belongsTo(Transaction::class,'account_book_id','id');
     }
 
 }
